O2Books - refactored how books are created, added command to get books by name, added command to get a book that lists all possible books.
Spells - added spell types for use with books and for future use with lessons, renamed some spells that existed but did not match what those spells do in the HP cannon, added a new super class for some basic object transfigurations

Wit-Sharpening Potion - added potion that improves book learning by double.

Houses - fixed bug with creating scoreboard when houses are not enabled, added name tag colors, added reset command, fixed a bug with syncPoints() that didn't handle if a team was not in the map.

Other:
https://github.com/Azami7/Ollivanders2/issues/63
https://github.com/Azami7/Ollivanders2/issues/90